# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/stores": {
      "post": {
        "operationId": "StoresController_create",
        "summary": "Create a new store",
        "description": "[Admin] Create a new store.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The record has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    },
                    {
                      "$ref": "#/components/schemas/Store"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Stores"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "StoresController_findMany",
        "summary": "Get all the stores by filtering them",
        "description": "[Admin] Get all the filtered stores.",
        "parameters": [
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "NIF",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Role (1: Admin, 2: Store, 3: Customer, 4: Unapproved Store, 5: Unconfirmed, 6: Guest)",
            "schema": {
              "enum": [
                1,
                2,
                3,
                4,
                5,
                6
              ],
              "type": "number"
            }
          },
          {
            "name": "approved",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "email-confirmed",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "The records have been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/User"
                      },
                      {
                        "$ref": "#/components/schemas/Store"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Stores"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/stores/pagination": {
      "get": {
        "operationId": "StoresController_findManyWithPagination",
        "summary": "Get all the stores by filtering them (with pagination)",
        "description": "[Admin] Get all the filtered stores (with pagination).",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Results/page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "NIF",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Role (1: Admin, 2: Store, 3: Customer, 4: Unapproved Store, 5: Unconfirmed, 6: Guest)",
            "schema": {
              "enum": [
                1,
                2,
                3,
                4,
                5,
                6
              ],
              "type": "number"
            }
          },
          {
            "name": "approved",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "email-confirmed",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "The records have been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "allOf": [
                              {
                                "$ref": "#/components/schemas/User"
                              },
                              {
                                "$ref": "#/components/schemas/Store"
                              }
                            ]
                          }
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Stores"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/stores/{id}": {
      "get": {
        "operationId": "StoresController_findOne",
        "summary": "Get a store by ID",
        "description": "[Admin] Get the specified store by its ID (store_id).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The record has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    },
                    {
                      "$ref": "#/components/schemas/Store"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Stores"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "StoresController_update",
        "summary": "Update a store.",
        "description": "[Admin] Update the specified store by its ID (store_id) with the specified data.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The record has been successfully updated."
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Stores"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "StoresController_remove",
        "summary": "Delete a store",
        "description": "[Admin] Delete the specified store by its ID (store_id).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "description": "Delete mode (soft or hard). Default: soft",
            "schema": {
              "enum": [
                "soft",
                "hard"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The record has been successfully deleted."
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Stores"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/stores/{id}/restore": {
      "patch": {
        "operationId": "StoresController_restore",
        "summary": "Restore a store",
        "description": "[Admin] Restore a soft deleted store specified  by its ID (store_id).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The record has been successfully restored."
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Stores"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Librería Santa Bárbara API",
    "description": "API documentation for Librería Santa Bárbara",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The id of the user"
          },
          "user_type_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the user type"
          },
          "role_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the role"
          },
          "email": {
            "type": "string",
            "example": "example@example.com",
            "description": "The email of the user"
          },
          "NIF": {
            "type": "string",
            "example": "12345678A",
            "description": "The NIF of the user"
          },
          "address": {
            "type": "string",
            "example": "Calle Princesa 12B",
            "description": "The address of the user"
          },
          "phone_number": {
            "type": "string",
            "example": "+34661122334",
            "description": "The phone number"
          },
          "email_confirmed": {
            "type": "boolean",
            "example": false,
            "description": "The email confirmation status of the user"
          },
          "hash": {
            "type": "string",
            "example": null,
            "description": "Hash used to verify email at registration"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "example": "2021-01-01",
            "description": "The date when the user was created"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "example": null,
            "description": "The date when the user has been updated"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string",
            "example": null,
            "description": "The date when the user has been (soft) deleted"
          }
        },
        "required": [
          "id",
          "user_type_id",
          "role_id",
          "email",
          "NIF",
          "address",
          "phone_number",
          "email_confirmed",
          "hash",
          "created_at",
          "updated_at",
          "deleted_at"
        ]
      },
...
      "Store": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The id of the store"
          },
          "user_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the user"
          },
          "approved": {
            "type": "boolean",
            "example": false,
            "description": "Store approval store by a admin"
          },
          "name": {
            "type": "string",
            "example": "Store name",
            "description": "The name of the store"
          },
          "last_activity": {
            "format": "date-time",
            "type": "string",
            "example": "2021-01-01",
            "description": "Date when the store has interactuated with the app"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "example": null,
            "description": "Date when the store has been updated"
          }
        },
        "required": [
          "id",
          "user_id",
          "approved",
          "name",
          "last_activity",
          "updated_at"
        ]
      },
      "CreateStoreDto": {
        "type": "object",
        "properties": {
          "role_id": {
            "type": "number",
            "example": 2,
            "description": "The id of the role that the user will have",
            "enum": [
              1,
              2,
              3,
              4,
              5,
              6
            ]
          },
          "password": {
            "type": "string",
            "example": "test",
            "description": "The password of the store.",
            "minLength": 8
          },
          "address": {
            "type": "string",
            "example": "Calle Princesa 12B",
            "description": "The address of the user",
            "maxLength": 300
          },
          "phone_number": {
            "type": "string",
            "example": "+34661122334",
            "description": "The phone number of the user",
            "maxLength": 15
          },
          "email_confirmed": {
            "type": "boolean",
            "example": false,
            "default": false,
            "description": "Whether the email of the user has been confirmed or not by the himself."
          },
          "hash": {
            "type": "string",
            "example": null,
            "default": null,
            "description": "Hash used to verify the email of the user at registration"
          },
          "email": {
            "type": "string",
            "example": "store@store.com",
            "description": "The email of the store. Should be unique among stores.",
            "maxLength": 300
          },
          "NIF": {
            "type": "string",
            "example": "12345678A",
            "description": "The NIF of the store. Should be alphanumeric and unique among stores.",
            "maxLength": 15
          },
          "name": {
            "type": "string",
            "example": "Librería San Pablo",
            "description": "The name of the store. Should be unique among stores.",
            "maxLength": 100
          },
          "approved": {
            "type": "boolean",
            "example": false,
            "default": false,
            "description": "Whether the store has been approved by an admin or not."
          }
        },
        "required": [
          "role_id",
          "password",
          "email",
          "name"
        ]
      },
      "UpdateStoreDto": {
        "type": "object",
        "properties": {
          "role_id": {
            "type": "number",
            "example": 2,
            "description": "The id of the role that the user will have",
            "enum": [
              1,
              2,
              3,
              4,
              5,
              6
            ]
          },
          "password": {
            "type": "string",
            "example": "test",
            "description": "The password of the store.",
            "minLength": 8
          },
          "address": {
            "type": "string",
            "example": "Calle Princesa 12B",
            "description": "The address of the user",
            "maxLength": 300
          },
          "phone_number": {
            "type": "string",
            "example": "+34661122334",
            "description": "The phone number of the user",
            "maxLength": 15
          },
          "email_confirmed": {
            "type": "boolean",
            "example": false,
            "default": false,
            "description": "Whether the email of the user has been confirmed or not by the himself."
          },
          "hash": {
            "type": "string",
            "example": null,
            "default": null,
            "description": "Hash used to verify the email of the user at registration"
          },
          "email": {
            "type": "string",
            "example": "store@store.com",
            "description": "The email of the store. Should be unique among stores.",
            "maxLength": 300
          },
          "NIF": {
            "type": "string",
            "example": "12345678A",
            "description": "The NIF of the store. Should be alphanumeric and unique among stores.",
            "maxLength": 15
          },
          "name": {
            "type": "string",
            "example": "Librería San Pablo",
            "description": "The name of the store. Should be unique among stores.",
            "maxLength": 100
          },
          "approved": {
            "type": "boolean",
            "example": false,
            "default": false,
            "description": "Whether the store has been approved by an admin or not."
          }
        }
      },
...
    }
  }
}